removed service bus and  modified other changes.
