+++ EventHubs/requirements.txt
@@ -4,6 +4,6 @@ azure-identity==1.15.0
 azure-core==1.29.6
 azure-mgmt-resource==23.0.1
 azure-mgmt-eventhub==11.0.0
-azure-servicebus==0.21.1
+azure-eventhub==5.11.5
 azure-mgmt-storage==21.1.0
 zure-mgmt-loganalytics==12.0.0
